pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-whois>=0.9.5",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.0",
]
